refactor(games-list): extract secondary-display/video into a part file
Move the secondary-display, video-preview, background and music-ducking
methods out of the `_SystemGamesListState` monolith into a new
`my_games_list/secondary_display.dart` part as `extension
_SecondaryDisplay`: `_resetVideoState`, `_stopVideoAndCleanup`,
`_performBackgroundOperationsForSelectedGame`, `_updateSecondaryDisplay`,
`_updateSecondaryDisplayVideo`, `_updateMusicDucking`, `_updateBackground`,
`_startVideoTimer`, `_initializeVideo`, `_getVideoPath`.

Mechanism (a) part/extension split — behaviour unchanged, no public-API
change. All state stays on the host; only methods move. `setState` calls
route through the host `rebuild()` bridge; the host statics `_log` and
`_videoDelay` are qualified as `_SystemGamesListState.*` (both required
from an extension).

Host 2763 -> 2290.
